+++ webapp/src/main/java/com/dentalcura/webapp/service/impl/UserService.java
@@ -6,6 +6,7 @@ import com.dentalcura.webapp.model.Patient;
 import com.dentalcura.webapp.model.User;
 import com.dentalcura.webapp.repository.IUserRepository;
 import com.dentalcura.webapp.service.IUserService;
+import com.dentalcura.webapp.utils.exceptions.CustomNotFoundException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.Getter;
 import lombok.Setter;
@@ -53,6 +54,9 @@ public class UserService implements IUserService {
 
     @Override
     public UserResponse selectUserByID(Long id) {
+        if (!userRepository.existsById(id))
+            throw new CustomNotFoundException("User id [" + id + "] not found");
+
         Optional<User> user = userRepository.findById(id);
         UserResponse userResponse = null;
 
@@ -64,6 +68,9 @@ public class UserService implements IUserService {
 
     @Override
     public void updateUserByID(Long id, UpdateUserRequest updateUserRequest) {
+        if (!userRepository.existsById(id))
+            throw new CustomNotFoundException("User id [" + id + "] not found");
+
         Optional<User> optionalUser = userRepository.findById(id);
 
         if (optionalUser.isPresent()) {
@@ -81,16 +88,13 @@ public class UserService implements IUserService {
             LOGGER.info("User [" + user.getName() + " " + user.getSurname() + "] updated");
         }
 
-
-//        LOGGER.info("Request to update user");
-//        User user = mapper.convertValue(updateUserRequest, User.class);
-//        user.setId(id);
-//        userRepository.save(user);
-//        LOGGER.info("User updated to [" + user.getName() + " " +user.getSurname() + "]");
     }
 
     @Override
     public void deleteUserByID(Long id) {
+        if (!userRepository.existsById(id))
+            throw new CustomNotFoundException("User id [" + id + "] not found");
+
         userRepository.deleteById(id);
         LOGGER.info("User deleted from DB");
     }
